pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "civStation"
version = "0.0.1"
requires-python = ">=3.10"
description = "Computer-use action evaluation framework for game environments. Collect data automatically and evaluate agent actions against ground truth."
readme = "README.md"
authors = [
    { name = "minsing-jin", email = "developerminsing@gmail.com" }
]
maintainers = [
    { name = "minsing-jin", email = "developerminsing@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Games/Entertainment :: Turn Based Strategy",
]
license = "MIT"
dependencies = [
    "anthropic>=0.77.0",
    "configargparse>=1.7.1",
    "dotenv>=0.9.9",
    "fastapi>=0.129.0",
    "google-genai>=1.60.0",
    "mcp>=1.12,<2",
    "mss>=9.0.0",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "pre-commit>=4.5.1",
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54",
    "pyyaml>=6.0.3",
    "rich>=13.0",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.5",
    "typer",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
# Voice input dependencies for HITL
voice = [
    "SpeechRecognition>=3.10.0",
    "PyAudio>=0.2.14",
    "openai-whisper>=20231117",  # For local Whisper STT
]
# Chat app integration dependencies
chatapp = [
    "discord.py>=2.3.0",
]
# Real-time status dashboard
ui = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]
docs = [
    "mkdocs-material>=9.6.0",
    "mkdocs-static-i18n>=1.3.1",
    "pymdown-extensions>=10.0",
]
test = [
    "coverage", # testing
    "pytest", # testing
    "ruff", # linting
    "ty", # checking types
    "ipdb", # debugging
    "pre-commit", # git hooks
]

[project.urls]
bugs = "https://github.com/minsing-jin/civStation/issues"
changelog = "https://github.com/minsing-jin/civStation/blob/main/changelog.md"
documentation = "https://minsing-jin.github.io/civStation/"
homepage = "https://github.com/minsing-jin/civStation"

[project.scripts]
civStation = "civStation.cli:app"
civStation_mcp = "civStation.mcp.server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --strict-markers"
markers = [
    "integration: marks tests requiring full environment (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["civStation"]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]

select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # Pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "UP", # pyupgrade
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["civStation*"]
